Refactor: move `CountDownTimer` to dashboard action page
