+++ src/modules/dashboard/components/ActionsPage/CountDownTimer/CountDownTimer.css
